change id type

+++ src/main/java/com/nrs/school/back/repository/StudentRepository.java
@@ -1,14 +1,14 @@
 package com.nrs.school.back.repository;
 
-import com.nrs.school.back.entities.Student;
-
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.nrs.school.back.entities.Student;
+
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
     Optional<Student> findByRegistration(String registration);
     void deleteByRegistration(String registration);
